backend/requirements.txt
# Flask核心框架
Flask==2.3.3
Flask-CORS==4.0.0

# JSON Web Token支持
PyJWT==2.8.0

# 密码加密
Werkzeug==2.3.7

# 数据库支持 (MySQL)
# SQLAlchemy==2.0.21
# Flask-SQLAlchemy==3.0.5
# PyMySQL==1.1.0

# API限流
# Flask-Limiter==3.5.0

# 日志和配置
python-dotenv==1.0.0

# 开发工具
pytest==7.4.2
pytest-flask==1.2.0

# 生产环境服务器
gunicorn==21.2.0

# 文件上传处理
# Pillow==10.0.1

# 时间处理
python-dateutil==2.8.2

# HTTP客户端 (用于与其他服务通信)
requests==2.31.0

# 数据验证
# marshmallow==3.20.1
# Flask-Marshmallow==0.15.0

# 缓存支持
# Flask-Caching==2.1.0

# 异步任务支持 (可选)
# Celery==5.3.1
# Redis==4.6.0

# 监控和指标 (可选)
# prometheus-client==0.17.1

# 配置管理
# Flask-Migrate==4.0.5